Adding ChaosDoor.cs to handle door mechanisms separately from room completion
