+++ Cosmic Quest - Order and Chaos/Assets/Animators/Rooms/Level1/ActivatedLetter.controller
@@ -9,7 +9,7 @@ AnimatorStateTransition:
   m_Name: 
   m_Conditions:
   - m_ConditionMode: 1
-    m_ConditionEvent: RoomCompleted
+    m_ConditionEvent: Reveal
     m_EventTreshold: 0
   m_DstStateMachine: {fileID: 0}
   m_DstState: {fileID: -2060031449026351040}
@@ -60,8 +60,8 @@ AnimatorController:
   m_Name: ActivatedLetter
   serializedVersion: 5
   m_AnimatorParameters:
-  - m_Name: RoomCompleted
-    m_Type: 4
+  - m_Name: Reveal
+    m_Type: 9
     m_DefaultFloat: 0
     m_DefaultInt: 0
     m_DefaultBool: 0
