+++ Cosmic Quest - Order and Chaos/Assets/Scripts/World/Level1/ChaosDoor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8ec0f8083076b7d428536ebbffda7217
+guid: 70f0a5c5c833c3f4ebc82d905db13730
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
