+++ Cosmic Quest - Order and Chaos/Assets/Scripts/World/Level1/Lvl1Room2.cs
@@ -1,0 +1,24 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Lvl1Room2 : Room
+{
+    protected AudioSource audioClip;
+    public Animator letterAnim;
+
+    private void Start()
+    {
+        GameObject letter = transform.Find("ActivatedLetter").gameObject;
+        audioClip = letter.GetComponent<AudioSource>();
+        letterAnim = letter.GetComponent<Animator>();
+    }
+
+    void Update()
+    {
+        if (AreAllEnemiesKilled())
+        {
+            letterAnim.SetTrigger("Reveal");
+        }
+    }
+}
