+++ Cosmic Quest - Order and Chaos/Assets/Scripts/World/Level1/Lvl1Room2.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8ec0f8083076b7d428536ebbffda7217
+guid: e7457271b814cca4787ef182e4b3675f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
