+++ Cosmic Quest - Order and Chaos/Assets/Scripts/World/Level1/Lvl1Room3.cs
@@ -4,13 +4,11 @@ using UnityEngine;
 
 public class Lvl1Room3 : Room
 {
+    protected AudioSource audioClip;
     public Animator letterReveal;
-    
+
     private void Start()
     {
-        // TODO: Implement random generator for lever code patterns based on input of code length and active player colours
-        code = new List<CharacterColour>();
-        input = new List<CharacterColour>();
         letterReveal = transform.parent.Find("ActivatedLetter").gameObject.GetComponent<Animator>();
     }
 
@@ -18,24 +16,7 @@ public class Lvl1Room3 : Room
     {
         if (AreAllEnemiesKilled())
         {
-
-        }
-    }
-
-    void OnTriggerEnter (Collider other) 
-    {
-        if (other.tag == "Player")
-        {
-            audioClip.Play(0);
-            Anim.SetTrigger("OpenDoor");
-        }
-    }
-
-    void OnTriggerExit(Collider other)
-    {
-        if (other.tag == "Player")
-        {
-            Anim.enabled = true;
+            letterReveal.SetTrigger("Reveal");
         }
     }
 }
